Display base RC panel
