+++ src/web/components/RemoteControlPanel/RemoteControlPanel.less
@@ -1,0 +1,9 @@
+@import "../../style/stylesheets/vars.less";
+
+.remote-control-panel {
+    position: fixed;
+    bottom: 0;
+    width: 100%;
+    height: 200px;
+    background-color: @overlay-background;
+}
